backend/pyproject.toml
[tool.poetry]
name = "qines-gai-backend"
version = "0.1.0"
description = "API Server for QINeS GAI"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
pdfplumber = "^0.10.3"
fastapi = "^0.111.0"
uvicorn = {version = "^0.24.0.post1", extras = ["standard"]}
httpx = "^0.27.0"
tqdm = "^4.66.1"
pypdf = "^4.2.0"
pdf2image = "^1.17.0"
lxml = "^5.3.0"
python-multipart = "^0.0.19"
asyncpg = "^0.30.0"
nanoid = "^2.0.0"
boto3 = "^1.35.85"
boto3-stubs = {extras = ["cognito-idp", "s3"], version = "^1.35.85"}
sqlalchemy = "^2.0.36"
meilisearch-python-sdk = "^4.0.0"
pytz = "^2024.2"
joserfc = "^1.0.2"
langchain = "^0.3.16"
langchain-openai = "^0.3.2"
langgraph = "^0.2.68"
python-pptx = "^1.0.2"
langchain_ollama = "^0.2.2"
langfuse = "^2.60.3"
markitdown = "^0.1.3"
langchain-community = "^0.3.0"
pydantic = ">=2.7.4,<3"
numpy = ">=1.26.4"
langchain-mcp-adapters = "^0.1.11"

[tool.poetry.group.dev.dependencies]
ruff = "^0.14.5"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-mock = "^3.14.0"
pytest-cov = "^6.0.0"
moto = {extras = ["cognito-idp", "s3"], version = "^5.0.24"}

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
fixable = ["I", "F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
pythonpath = "."
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
